+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # v0.3.0
 
 * Adds new API methods `RotateAPIKey` and `CheckPermission`.
-* Provide API methods that return secret data as an `io.ReadCloser` rather than of `[]byte`. This way, the API client gets the only copy of the secret data and can handle it however she sees fit.
+* Provides API methods that return secret data as an `io.ReadCloser` rather than of `[]byte`. This way, the API client gets the only copy of the secret data and can handle it however she sees fit.
 * Loading a policy requires `PolicyMode` argument.
+* Loading a policy returns `PolicyResponse`. 
 
 # v0.2.0